Make create.yml resistant to errors after VM clone
Previously if we had any failure (for instance proxmox internal failure)
we didn't remove the cloned VM, which (depending on the stage) resulted
in left over VMs (running or just cloned) left on the proxmox instance.

Don't use proxmox_kvm ansible module - just rely on the direct calls to
proxmox. This happens to be more reliable in our case, giving better
control for cases like attempt to delete already deleted VM. For us it's
better to get the response from proxmox and handle it the way we won't
instead of getting the module status which may hide some details.